projects/datasets/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "birdsys-datasets"
version = "0.1.0"
description = "BirdSys dataset building workflows"
requires-python = ">=3.11,<3.12"
dependencies = [
  "duckdb>=1.1.3,<2",
  "pandas>=2.2,<3",
  "pillow>=10,<12",
  "pyarrow>=18,<20",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["birdsys.datasets*"]
namespaces = true
